Moved classes from beautifulSoupSelenium.py to respective files. Added more detailed progress indications.
